+++ CHANGELOG.md
@@ -23,6 +23,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Fixed
 - Avoid QueryPlan crash when multiple $fieldNodes are present
 - Clarify error when attempting to coerce anything but `array` or `stdClass` to an input object
+- Allow directives on variable definitions
 
 ### Removed
 - Removed deprecated `Type::getInternalTypes()`