+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Added
 
 - Improve extendability of validator rules
+- Add tests for errors that occur when undeclared fields are passed in input
 
 ### Optimized
 
@@ -27,6 +28,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Avoid QueryPlan crash when multiple $fieldNodes are present
 - Clarify error when attempting to coerce anything but `array` or `stdClass` to an input object
+- Handle `null` in `Type::getNullableType`
 
 ### Removed