+++ CHANGELOG.md
@@ -28,7 +28,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 
 - Avoid QueryPlan crash when multiple $fieldNodes are present
 - Clarify error when attempting to coerce anything but `array` or `stdClass` to an input object
-- Handle `null` in `Type::getNullableType`
+- Handle `null` parent of list in `ValuesOfCorrectType::getVisitor`
 
 ### Removed